requirements.txt
#certifi==2019.11.28
#chardet==3.0.4
Django==1.10.4
djangorestframework==3.5.3
get==2019.4.13
idna==2.8
post==2019.4.13
public==2019.4.13
query-string==2019.4.13
#request==2019.4.13
requests==2.22.0
azure-cosmos==3.1.2

